[bugfix]将参数类型转换成u16

+++ Core/Inc/ui.h
@@ -41,7 +41,7 @@ typedef struct PageNodeTypedef
 
 } PageNodeTypedef;
 
-uint8_t item_node_text_init(ItemNodeTypedef *NodeItem, uint8_t text_loction_x, uint8_t text_loction_y);
+uint8_t item_node_text_init(ItemNodeTypedef *NodeItem, uint16_t text_loction_x, uint16_t text_loction_y);
 void page_node_init(PageNodeTypedef *NodePage);
 void curve_animotion(void);
 void show_UI(void);
